requirements.txt
fastapi
uvicorn
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
psycopg2-binary
sqlalchemy
python-dotenv
pydantic
python-multipart
slowapi
email-validator